Rename com.google.cloud.tools.intellij.git package to com.google.cloud.tools.intellij.vcs
'git' is not the higher order concept in the source repo integration.
