+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/vcs/ChooseProjectDialog.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.git;
+package com.google.cloud.tools.intellij.vcs;
 
 import com.google.cloud.tools.intellij.elysium.ProjectSelector;
 import com.google.cloud.tools.intellij.login.CredentialedUser;
