+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/vcs/GcpHttpAuthDataProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.git;
+package com.google.cloud.tools.intellij.vcs;
 
 import com.google.api.client.repackaged.com.google.common.base.Strings;
 import com.google.cloud.tools.intellij.elysium.SelectUserDialog;
