+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/vcs/GcpHttpAuthProviderTest.java.excluded
